fix(app,tests): headless-safe AsyncCommand error dialog + clear CodeQL test-code alerts
Root cause of the intermittent 6h CI build-test timeout: AsyncCommand.TryShowError
called MessageBox.Show on every caught exception in ExecuteAsync. On a headless CI
runner (no message pump to dismiss it) the modal blocked indefinitely, hitting the
6-hour job ceiling. The AsyncCommand_*_Catches wave-9 tests deliberately trigger
these catch blocks, so the suite hung intermittently. Now skip the dialog when no
WPF Application is running (test / app-shutdown contexts).

Also clear the CodeQL alerts the diff-informed scan flagged on this branch's new
test code so the PR does not add to code-scanning debt:
- Path.Combine -> Path.Join (silent earlier-argument drop)
- remove useless upcasts and redundant same-type casts
- remove useless local-variable assignments
- catch specific IOException/UnauthorizedAccessException instead of catch-all in cleanup
- null-forgiving access on possibly-null dereferences

Full deterministic suite: 5,358 passed, 0 failed, no hang dump.

+++ tests/SwfocTrainer.Tests/Runtime/SignatureResolverWave8Tests.cs
@@ -327,7 +327,7 @@ public sealed class SignatureResolverWave8Tests
     public void TryParseAddress_IntValue_ReturnsTrue()
     {
         var method = typeof(SignatureResolverSymbolHydration).GetMethod("TryParseAddress", NonPublicStatic)!;
-        var args = new object?[] { (int)0x1234, 0L };
+        var args = new object?[] { 0x1234, 0L };
         var result = (bool)method.Invoke(null, args)!;
         result.Should().BeTrue();
         ((long)args[1]!).Should().Be(0x1234);
